refactor MagiclinksCtr
